+++ src/SVM.cpp
@@ -40,7 +40,7 @@ cl_int SharedVirtualMemory::getHostOffset(const void* hostPointer) const
 
 void* SharedVirtualMemory::getDevicePointer(const size_t offset)
 {
-	return reinterpret_cast<void*>(buffer->qpuPointer + offset);
+	return reinterpret_cast<void*>(static_cast<uint32_t>(buffer->qpuPointer) + offset);
 }
 
 void* SharedVirtualMemory::getHostPointer(const size_t offset)
